requirements.txt
pycryptodome
matplotlib
hexdump